<fix>[storage]: fix mark vol as snapshot rollback
handle NPE

Resolves: ZSTAC-63723



(cherry picked from commit 633b9ce38b27423ab4165b25cfd6825111d4800d)

+++ storage/src/main/java/org/zstack/storage/snapshot/VolumeSnapshotManagerImpl.java
@@ -675,6 +675,9 @@ public class VolumeSnapshotManagerImpl extends AbstractService implements
     @Transactional
     private void rollbackSnapshot(String uuid) {
         VolumeSnapshotVO vo = dbf.getEntityManager().find(VolumeSnapshotVO.class, uuid);
+        if (vo == null) {
+            return;
+        }
 
         dbf.getEntityManager().remove(vo);
